feat: widen block-id storage to sixteen bits (#17)
The kernel's ceiling has been sixteen bits for some time; this package
stored block ids one byte per cell and carried its own maximum as a
hardcoded byte-wide literal that did not track the kernel's. Past two
hundred and fifty-six blocks, ids would wrap silently and the mesher
would build wrong geometry for the wrapped ones, with nothing raised.

This is the work a previous change deliberately deferred rather than
overlooked. When the kernel pin moved, its changeset recorded that
re-exporting the kernel's wider ceiling into a byte-wide mask would
corrupt the greedy mesher's packed cell, and kept the old literal as a
stopgap. That reasoning was right, and it is why this change is more
than a type swap.

The greedy mask packs a block id together with ambient-occlusion bits
in one cell and merges while the whole cell repeats. Its shift sat
directly above the old byte-wide id, so widening the storage alone
would have reintroduced exactly the corruption that stopgap avoided,
just at a wider boundary — any id above sixteen thousand three hundred
and eighty-three colliding with the occlusion bits. The shift moves to
sixteen and every buffer backing that mask widens again, because an id
of sixteen bits plus two occlusion bits no longer fits in sixteen.

The packed vertex output carried a second, independent truncation of
the same kind and is widened with it.

The maximum is now re-exported from the kernel rather than restated,
so the two cannot drift apart again — which was the original defect.
Lookup tables keyed by block id were already sized from that constant
and grow on their own; their element types are unchanged. Light grids,
fluid state and rail shapes are deliberately untouched: those really
are one byte per cell and have nothing to do with block ids.

No persisted or wire format exists here — the view is purely in-memory
and the caller copies opaque kernel storage in at the boundary — so
unlike the generator's equivalent change there is no version bump and
no migration to invent.

The existing occlusion regression was pinned at the old byte ceiling
and now uses the real one, which byte storage cannot represent, and a
new test meshes an id that could only ever have been stored as
forty-four and asserts it survives. Two assertions that hardcoded the
old table length were failing against the widened table and are
corrected — they had gone stale rather than staying meaningful.

+++ src/domain/chunk-view.ts
@@ -2,7 +2,7 @@
  * The chunk as meshing sees it, and the hot-path block read.
  *
  * `ChunkView` is the renderer-facing variable-height view. It keeps the
- * mc-kernel coordinate and height while carrying the byte view used by the
+ * mc-kernel coordinate and height while carrying the element view used by the
  * meshing hot path. The caller copies opaque kernel block storage at the
  * boundary; this package does not expose a second adapter API.
  */
@@ -112,7 +112,7 @@ export type ChunkView = Omit<KernelChunk, 'blocks' | 'height'> & {
   /** Vertical extent of this chunk, in blocks. */
   readonly height: number
   /** `CHUNK_SIZE * height * CHUNK_SIZE` block ids, laid out per `blockIndex`. */
-  readonly blocks: Readonly<Uint8Array>
+  readonly blocks: Readonly<Uint16Array>
   /**
    * Per-cell decoded fluid state when available from the simulation boundary.
    *
@@ -134,7 +134,7 @@ export const BLOCKS_PER_CHUNK: number = blockCountOf(CHUNK_HEIGHT)
 
 /** An all-air chunk. Useful as a neighbour for an edge chunk, and in tests. */
 export const emptyChunk = (height: number = CHUNK_HEIGHT): ChunkView => ({
-  blocks: new Uint8Array(blockCountOf(height)),
+  blocks: new Uint16Array(blockCountOf(height)),
   coord: chunkCoord(ORIGIN_CHUNK_AXIS, ORIGIN_CHUNK_AXIS),
   height,
 })
